Move prescribing conversion to convert_ command
