+++ openprescribing/frontend/tests/fixtures/commands/detailed_prescribing.csv
@@ -1,4 +1,5 @@
 Regional Office Name,Regional Office Code,Area Team Name,Area Team Code,PCO Name,PCO Code,Practice Name,Practice Code,BNF Code,BNF Description,Items,Quantity,ADQ Usage,NIC,Actual Cost
+NORTH OF ENGLAND,Y54,GREATER MANCHESTER AREA TEAM,Q46,WIGAN BOROUGH CCG,02H00,DR CHAN,P92042,0202010B0AAABAB,Bendroflumethiazide_Tab 2.5mg,46,14,644.00000,15.41,14.55976
 NORTH OF ENGLAND,Y54,GREATER MANCHESTER AREA TEAM,Q46,WIGAN BOROUGH CCG,02H00,DR CHAN,P92042,0202010B0AAABAB,Bendroflumethiazide_Tab 2.5mg,23,28,644.00000,15.41,14.55976
 NORTH OF ENGLAND,Y54,GREATER MANCHESTER AREA TEAM,Q46,WIGAN BOROUGH CCG,02H00,DR CHAN,P92042,0302000N0AABFBF,Fluticasone/Salmeterol_Inh 125/25mcg120D,2,1,60.00000,70.00,64.86705
 NORTH OF ENGLAND,Y54,MERSEYSIDE AREA TEAM,Q48,SOUTHPORT AND FORMBY CCG,01V00,CURZON ROAD MEDICAL PRACTICE,N84013,0106020M0AAATAT,Senna_Oral Soln 7.5mg/5ml S/F,1,280,28.00000,2.67,2.58567
@@ -8,3 +9,4 @@ NORTH OF ENGLAND,Y54,NORTH YORKSHIRE AND HUMBER AREA TEAM,Q50,EAST RIDING OF YOR
 NORTH OF ENGLAND,Y54,WEST YORKSHIRE AREA TEAM,Q52,WAKEFIELD CCG,03R00,STUART ROAD,B87015,21300000100,Oxyal Sod Hyaluronate Eye Dps 10ml,1,2,0.00000,8.30,7.70084
 NORTH OF ENGLAND,Y54,"CUMBRIA,NORTHUMB,TYNE & WEAR AREA TEAM",Q49,CUMBRIA CCG,01H00,DALSTON MEDICAL GROUP,A82022,1104020N0AAAAAA,Nedocromil Sod_Eye Dps Aq 2%,1,10,0.00000,5.72,5.31094
 MIDLANDS AND EAST OF ENGLAND,Y55,HERTFORDSHIRE & SOUTH MIDLANDS AREA TEAM,Q58,MILTON KEYNES CCG,04F00,MILTON KEYNES VILLAGE SURG,K82631,0906040G0BXAAAN,Fultium-D3_Cap 800u,13,30,0.00000,46.15,43.01077
+MIDLANDS AND EAST OF ENGLAND,Y55,HERTFORDSHIRE & SOUTH MIDLANDS AREA TEAM,Q58,MILTON KEYNES CCG,04F00,UNKNOWN SURGERY,Q58998,0906040G0BXAAAN,Fultium-D3_Cap 800u,13,30,0.00000,46.15,43.01077
